app/requirements.txt
fastapi==0.87.0
uvicorn[standard]==0.19.0
sqlalchemy==1.4.46
asyncpg==0.27.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
aiosmtplib==2.0.1
aioredis==2.0.1
email-validator==1.3.1
